Update banlist
To avoid getting the error "[LUA ERROR] playerAuthHandler: [string "Resources\Server\banman\main.lua"]:25: attempt to index a nil value (global 'banfile')", there needs to be a user in the ban list. This user has been made from a password generator and therefore is very unlikely to be anyone's username.